refactoring and bugfix
